Split predict.py into train and predict
